+++ airborn.web/Models/Aircraft/SR22.cs
@@ -16,30 +16,30 @@ namespace Airborn.web.Models
             PerformanceCalculationLogItem logItem)
         {
 
-            decimal adjustedTakeoffDistance = unadjustedTakeoffDistance.TotalFeet;
+            double adjustedTakeoffDistance = unadjustedTakeoffDistance.TotalFeet;
 
             if (result.HeadwindComponent > 0)
             {
 
                 // subtract 10% for each 12 knots of headwind
-                adjustedTakeoffDistance = adjustedTakeoffDistance * (1 - ((result.HeadwindComponent / 12) * 0.1m));
+                adjustedTakeoffDistance = adjustedTakeoffDistance * (1 - ((result.HeadwindComponent / 12) * 0.1f));
 
                 logItem.Add("Subtracting 10% for each 12 knots of headwind");
                 logItem.Add("Headwind component: " + result.HeadwindComponent.ToString("#,##0.00") + " kts");
                 logItem.Add("Unadjusted takeoff distance: " + unadjustedTakeoffDistance.TotalFeet.ToString("#,##0.00") + " ft");
-                logItem.Add("Adjustment factor (headwind component / 12 * 10%): " + ((result.HeadwindComponent / 12) * 0.1m).ToString("#,##0.00"));
+                logItem.Add("Adjustment factor (headwind component / 12 * 10%): " + ((result.HeadwindComponent / 12) * 0.1f).ToString("#,##0.00"));
             }
             else if (result.HeadwindComponent < 0) // negative headwinds are tailwinds
             {
 
                 // add 10% for each 2 knots of tailwind up to 10 knots
-                decimal adjustment = (result.HeadwindComponent / 2) * 0.1m;
+                double adjustment = (result.HeadwindComponent / 2) * 0.1f;
 
                 // don't allow more than 50% adjustment
-                if (adjustment > 0.5m)
+                if (adjustment > 0.5f)
                 {
                     logItem.Add("Adjustment capped at 50%");
-                    adjustment = 0.5m;
+                    adjustment = 0.5f;
 
                 }
 
@@ -49,14 +49,14 @@ namespace Airborn.web.Models
                 logItem.Add("Adding 10% for each 2 knots of tailwind up to 10 knots");
                 logItem.Add("Tailwind component: " + result.HeadwindComponent.ToString("#,##0.00") + " kts");
                 logItem.Add("Unadjusted takeoff distance: " + unadjustedTakeoffDistance.TotalFeet.ToString("#,##0.00") + " ft");
-                logItem.Add("Adjustment factor (tailwind component / 2 * 10%): " + ((result.HeadwindComponent / 2) * 0.1m).ToString("#,##0.00"));
+                logItem.Add("Adjustment factor (tailwind component / 2 * 10%): " + ((result.HeadwindComponent / 2) * 0.1f).ToString("#,##0.00"));
 
             }
 
             if (result.Runway.Slope.HasValue && result.Runway.Slope > 0)
             {
                 logItem.Add("Runway slope: " + result.Runway.Slope.Value.ToString("0.00") + " %");
-                logItem.Add("Runway slope adjustment factor: " + (result.Runway.Slope.Value * 0.1m).ToString("0.00"));
+                logItem.Add("Runway slope adjustment factor: " + (result.Runway.Slope.Value * 0.1f).ToString("0.00"));
 
                 adjustedTakeoffDistance =
                     CalculateAdjustedGroundRollDistanceForTakeoff(
@@ -92,7 +92,7 @@ namespace Airborn.web.Models
             double adjustment = Math.Abs(slope) * adjustmentFactor * groundRollDistance;
             double adjustedGroundRollDistance = groundRollDistance + adjustment;
 
-            return Distance.FromFeet((decimal)adjustedGroundRollDistance);
+            return Distance.FromFeet((double)adjustedGroundRollDistance);
         }
 
         public override Distance MakeLandingAdjustments(
@@ -100,29 +100,29 @@ namespace Airborn.web.Models
             Distance unadjustedLandingDistance,
             PerformanceCalculationLogItem logItem)
         {
-            decimal adjustedLandingDistance = unadjustedLandingDistance.TotalFeet;
+            double adjustedLandingDistance = unadjustedLandingDistance.TotalFeet;
 
             if (result.HeadwindComponent > 0)
             {
                 // subtract 10% for each 13 knots of headwind
-                adjustedLandingDistance = adjustedLandingDistance * (1 - ((result.HeadwindComponent / 13) * 0.1m));
+                adjustedLandingDistance = adjustedLandingDistance * (1 - ((result.HeadwindComponent / 13) * 0.1f));
 
                 logItem.Add("Subtracting 10% for each 13 knots of headwind");
                 logItem.Add("Headwind component: " + result.HeadwindComponent.ToString("#,##0.00") + " kts");
                 logItem.Add("Unadjusted landing distance: " + unadjustedLandingDistance.TotalFeet.ToString("#,##0.00") + " ft");
-                logItem.Add("Adjustment factor (headwind component / 13 * 10%): " + ((result.HeadwindComponent / 13) * 0.1m).ToString("#,##0.00"));
+                logItem.Add("Adjustment factor (headwind component / 13 * 10%): " + ((result.HeadwindComponent / 13) * 0.1f).ToString("#,##0.00"));
             }
             else if (result.HeadwindComponent < 0)
             {
                 // negative headwinds are tailwinds
 
                 // add 10% for each 2 knots of tailwind up to 10 knots
-                decimal adjustment = (result.HeadwindComponent / 2) * 0.1m;
+                double adjustment = (result.HeadwindComponent / 2) * 0.1f;
 
                 // don't allow more than 50% adjustment
-                if (adjustment > 0.5m)
+                if (adjustment > 0.5f)
                 {
-                    adjustment = 0.5m;
+                    adjustment = 0.5f;
                     logItem.Add("Adjustment capped at 50%: " + adjustment.ToString("#,##0.00"));
                 }
 
@@ -139,7 +139,7 @@ namespace Airborn.web.Models
             if (result.Runway.Slope.HasValue && result.Runway.Slope > 0)
             {
                 logItem.Add("Runway slope: " + result.Runway.Slope.Value.ToString("0.00") + " %");
-                logItem.Add("Runway slope adjustment factor: " + (result.Runway.Slope.Value * 0.1m).ToString("0.00"));
+                logItem.Add("Runway slope adjustment factor: " + (result.Runway.Slope.Value * 0.1f).ToString("0.00"));
 
                 adjustedLandingDistance =
                     CalculateAdjustedGroundRollDistanceForLanding(
@@ -200,7 +200,7 @@ namespace Airborn.web.Models
             double adjustedGroundRollDistance = groundRollDistance + adjustment;
 
             // Return the adjusted ground roll distance
-            return Distance.FromFeet((decimal)adjustedGroundRollDistance);
+            return Distance.FromFeet((double)adjustedGroundRollDistance);
         }
 
     }
