+++ airborn.web/Models/PerformanceCalculations/Calculation/CalculationUtilities.cs
@@ -14,12 +14,12 @@ namespace Airborn.web.Models
         /// If you give it 0 and 90, it will return you 90
         /// If you give it 0 and 270, it will return you -90
         /// </summary>
-        public static decimal SmallestAngularDifference(int angle1, int angle2)
+        public static double SmallestAngularDifference(int angle1, int angle2)
         {
             // the smallest angular difference between two angles is the difference between the two angles
             // modulo 360, minus 180
 
-            decimal difference = (angle2 - angle1 + 180) % 360 - 180;
+            double difference = (angle2 - angle1 + 180) % 360 - 180;
 
             // if the difference is greater than 180, subtract 360 from it so that we never return a value greater than 180
             difference = difference < -180 ? difference + 360 : difference;
@@ -38,7 +38,7 @@ namespace Airborn.web.Models
         /// </summary>
         /// <param name="runwayHeading">The heading of the runway</param>
         /// <param name="windDirection">The direction of the wind</param>
-        public static decimal AngularDifferenceBetweenRunwayAndWind(int runwayHeading, int windDirection)
+        public static double AngularDifferenceBetweenRunwayAndWind(int runwayHeading, int windDirection)
         {
             // the angular difference between the runway and the wind is the smallest angular difference between the two
             return SmallestAngularDifference(runwayHeading, windDirection);
@@ -48,14 +48,14 @@ namespace Airborn.web.Models
         /// Converts degrees to radians
         /// </summary>
         /// <param name="degrees">The number of degrees to convert</param>
-        public static decimal DegreesToRadians(decimal degrees)
+        public static double DegreesToRadians(double degrees)
         {
             // 180 degrees = pi radians
             // 1 degree = pi / 180 radians
             // so, x degrees = (pi / 180) * x radians
             // or, in this case , x radians = (180 / pi) * x degrees
 
-            decimal radians = ((decimal)Math.PI / 180) * degrees;
+            double radians = ((double)Math.PI / 180) * degrees;
             return radians;
         }
 
@@ -63,7 +63,7 @@ namespace Airborn.web.Models
         /// Converts a temperature in fahrenheit to celcius
         /// </summary>
         /// <param name="fahrenheitTemperature">The temperature in fahrenheit to convert</param>
-        public static decimal FahrenheitToCelcius(decimal fahrenheitTemperature)
+        public static double FahrenheitToCelcius(double fahrenheitTemperature)
         {
             return ((fahrenheitTemperature - 32) * 5 / 9);
         }
@@ -72,10 +72,10 @@ namespace Airborn.web.Models
         /// Converts inches of mercury to millibars
         /// </summary>
         /// <param name="inchesOfMercury">The pressure in inches of mercury to convert</param>
-        public static decimal InchesOfMercuryToMillibars(decimal inchesOfMercury)
+        public static double InchesOfMercuryToMillibars(double inchesOfMercury)
         {
             // 1 inch of mercury = 33.8653074866 millibars
-            return inchesOfMercury * 33.8653074866m;
+            return inchesOfMercury * 33.8653074866f;
         }
 
 
@@ -83,12 +83,12 @@ namespace Airborn.web.Models
         /// Returns the ISA temperature at a given pressure altitude
         /// </summary>
         /// <param name="pressureAltitude">The pressure altitude to calculate the ISA temperature at</param>
-        public static decimal ISATemperatureForPressureAltitude(decimal pressureAltitude)
+        public static double ISATemperatureForPressureAltitude(double pressureAltitude)
         {
 
-            decimal isaTemperatureAtSeaLevel = 15; // 15 degrees Celsius
-            decimal temperatureLapseRate = 2; // 2 degrees Celsius per 1000 feet
-            decimal altitudeInterval = 1000;
+            double isaTemperatureAtSeaLevel = 15; // 15 degrees Celsius
+            double temperatureLapseRate = 2; // 2 degrees Celsius per 1000 feet
+            double altitudeInterval = 1000;
 
             // ISA temperature starts at 15 degrees Celsius at sea level, 
             // and decreases by 2 degrees Celsius for every 1000 feet of altitude
@@ -100,11 +100,11 @@ namespace Airborn.web.Models
         /// </summary>
         /// <param name="windStrength">The strength of the wind</param>
         /// <param name="windRunwayAngularDifferenceMagnetic">The angular difference between the runway and the wind</param>
-        public static decimal HeadwindComponent(int windStrength, decimal windRunwayAngularDifferenceMagnetic)
+        public static double HeadwindComponent(int windStrength, double windRunwayAngularDifferenceMagnetic)
         {
             // headwind component = wind strength * cos(the angular difference between the runway and the wind)
 
-            return (decimal)(
+            return (double)(
                 windStrength * Math.Cos((double)DegreesToRadians(windRunwayAngularDifferenceMagnetic))
             );
         }
@@ -115,11 +115,11 @@ namespace Airborn.web.Models
         /// </summary>
         /// <param name="windStrength">The strength of the wind</param>
         /// <param name="windRunwayAngularDifferenceMagnetic">The angular difference between the runway and the wind</param>
-        public static decimal CrosswindComponent(int windStrength, decimal windRunwayAngularDifferenceMagnetic)
+        public static double CrosswindComponent(int windStrength, double windRunwayAngularDifferenceMagnetic)
         {
             // crosswind component = wind strength * sin(the angular difference between the runway and the wind)
 
-            return Convert.ToDecimal(
+            return Convert.ToDouble(
                  windStrength * Math.Sin(
                     (double)CalculationUtilities.DegreesToRadians(windRunwayAngularDifferenceMagnetic)
                     )
@@ -131,14 +131,14 @@ namespace Airborn.web.Models
         /// </summary>
         /// <param name="QNH">The QNH at the airport</param>
         /// <param name="fieldElevation">The elevation of the airport</param>
-        public static decimal PressureAltitudeAtFieldElevation(decimal QNH, int fieldElevation)
+        public static double PressureAltitudeAtFieldElevation(double QNH, int fieldElevation)
         {
-            decimal standardPressure = 1013.25m;
+            double standardPressure = 1013.25f;
 
             // folks usually round this to 30, but, fun fact, the formula is actually
             // 96 × ( T in kelvin) / QNH (in hPa)
             // at T = ISA (15C) and QNH = 1013.25 hPa, this is 27.3
-            decimal lapseRate = 27.3m;
+            double lapseRate = 27.3f;
 
             return ((standardPressure - QNH) * lapseRate) + fieldElevation;
         }
@@ -149,13 +149,13 @@ namespace Airborn.web.Models
         /// <param name="temperatureCelcius">The temperature at the airport in celcius</param>
         /// <param name="isaTemperature">The ISA temperature at the airport</param>
         /// <param name="pressureAltitude">The pressure altitude at the airport</param>
-        public static decimal DensityAltitudeAtAirport(decimal temperatureCelcius, decimal isaTemperature, decimal pressureAltitude)
+        public static double DensityAltitudeAtAirport(double temperatureCelcius, double isaTemperature, double pressureAltitude)
         {
             // density altitude is the pressure altitude adjusted for temperature
             // the formula is:
             // density altitude = pressure altitude + (120 * (temperature - ISA temperature))
 
-            decimal densityAltitudeInterval = 120;
+            double densityAltitudeInterval = 120;
 
             return ((temperatureCelcius - isaTemperature) * densityAltitudeInterval) + pressureAltitude;
         }
